Tests reorganized and refactored
