+++ tests/Feature/PlayerTest.php
@@ -15,6 +15,75 @@ class PlayerTest extends TestCase
 {
     use Withfaker;
 
+        /** @test */
+        public function new_player_can_be_registered_by_admin_role() {
+
+            $this->withoutExceptionHandling();
+    
+            $userAdmin = User::factory()->create();
+            $userAdmin->assignRole(Role::findByName('admin', 'api'));
+            $userAdminToken = $userAdmin->createToken('admin')->accessToken;
+    
+            do {
+                $fakeName = $this->faker->firstName;
+            } while (strlen($fakeName <5));
+            do {
+                $fakeSurname = $this->faker->lastName;
+            } while (strlen($fakeSurname <5));
+            $fakeEmail = $this->faker->safeEmail;
+            $fakeHeight = rand(160, 200);
+            $fakeHand = (rand(1,2) == 1) ? 'left' : 'right';
+            $fakeBackhand = (rand(1,2) == 1) ? 'one hand' : 'two hands';
+            $fakeBriefing = $this->faker->sentence;
+    
+            $response = $this->post('/api/register_player',
+            [
+                'name' => $fakeName,
+                'surname' => $fakeSurname,
+                'email' => $fakeEmail,
+                'password' => bcrypt('xxx'),
+                'height' => $fakeHeight,
+                'playing_hand' => $fakeHand,
+                'backhand_style' => $fakeBackhand,
+                'briefing' => $fakeBriefing,
+    
+            ],
+            [
+                'Authorization' => 'Bearer ' . $userAdminToken
+            ]);
+    
+            $response->assertOk();
+    
+            $responseJson = $response->json();
+            $this->assertEquals($responseJson['response_code'], '200');
+            $this->assertEquals($responseJson['message'], 'Player registration successful');
+    
+            $player = Player::with('user')
+            ->get()
+            ->last();
+    
+            $playerInfo = [
+                'name' => $player['user']['name'],
+                'surname' => $player['user']['surname'],
+                'email' => $player['user']['email'],
+                'height' => $player['height'],
+                'playing_hand' => $player['playing_hand'],
+                'backhand_style' => $player['backhand_style'],
+                'briefing' => $player['briefing'],
+            ];
+    
+            $this->assertEquals($playerInfo['name'], $fakeName);
+            $this->assertEquals($playerInfo['surname'], $fakeSurname);
+            $this->assertEquals($playerInfo['email'], $fakeEmail);
+            $this->assertEquals($playerInfo['height'], $fakeHeight);
+            $this->assertEquals($playerInfo['playing_hand'], $fakeHand);
+            $this->assertEquals($playerInfo['backhand_style'], $fakeBackhand);
+            $this->assertEquals($playerInfo['briefing'], $fakeBriefing);
+    
+            $userAdmin->delete();
+            $player->delete();
+        }
+
      /** @test */
      public function can_get_To5Player_list() {
 
@@ -48,8 +117,9 @@ class PlayerTest extends TestCase
         $user = User::factory()->create();
         $user->assignRole(Role::findByName('user', 'api'));
         $userToken = $user->createToken('user')->accessToken;
+
         $player = Player::factory()->create();
-        $user = User::where('id', $player->user_id)->first();
+        $userPlayer = User::where('id', $player->user_id)->first();
 
         $response = $this->get("api/player_info/{$player->ranking}", 
             ['Authorization' => 'Bearer ' . $userToken]
@@ -59,8 +129,8 @@ class PlayerTest extends TestCase
 
         $responseJson = $response->json();
         $this->assertEquals($responseJson['info']['ranking'], $player->ranking);
-        $this->assertEquals($responseJson['info']['name'], $user->name);
-        $this->assertEquals($responseJson['info']['surname'], $user->surname);
+        $this->assertEquals($responseJson['info']['name'], $userPlayer->name);
+        $this->assertEquals($responseJson['info']['surname'], $userPlayer->surname);
         $this->assertEquals($responseJson['info']['height'], $player->height);
         $this->assertEquals($responseJson['info']['playing_hand'], $player->playing_hand);
         $this->assertEquals($responseJson['info']['backhand_style'], $player->backhand_style);
@@ -69,6 +139,7 @@ class PlayerTest extends TestCase
         $this->assertEquals($responseJson['message'], 'Player info delivered');
 
         $user->delete();
+        $userPlayer->delete();
 
     }
